+++ client/util/util.dart
@@ -11,6 +11,7 @@
 #import("../event/event.dart");
 
 #source("src/StringUtil.dart");
+#source("src/XMLUtil.dart");
 #source("src/ListUtil.dart");
 #source("src/MapUtil.dart");
 #source("src/AbstractList.dart");